+++ source/backend/Risk.API/Models/Rol.cs
@@ -28,7 +28,7 @@ namespace Risk.API.Models
 {
     public class Rol : IModel
     {
-        public int IdRol { get; set; }
+        public long IdRol { get; set; }
         public string Nombre { get; set; }
         public string Activo { get; set; }
         public string Detalle { get; set; }
